+++ R/quote.R
@@ -1,18 +1,29 @@
-#' Enumerate Multiple Strings
+#' Enumerate Multiple Elements
 #'
-#' @param x A `character` vector
-#' @param quote_fun Quoting function, defaults to `backquote`.
-#' @param conjunction Character to be used in the message, defaults to "and".
+#' Enumerate multiple elements of a vector or list.
 #'
-#' @author Thomas Neitmann
+#' @param x A vector or list
+#' @param quote_fun Quoting function, defaults to `backquote`. If set to `NULL`,
+#'   the elements are not quoted.
+#' @param conjunction Character to be used in the message, defaults to `"and"`.
+#'
+#' @author Thomas Neitmann, Stefan Bundfuss
 #'
 #' @return A `character` vector
 #'
 #' @keywords quote
 #' @family quote
 #'
+#' @examples
+#' enumerate(c("one", "two", "three"))
+#'
+#' enumerate(c(1, 2, 3), quote_fun = NULL)
+#'
 #' @export
 enumerate <- function(x, quote_fun = backquote, conjunction = "and") {
+  if (is.null(quote_fun)) {
+    quote_fun <- function(x) x
+  }
   if (length(x) == 1L) {
     quote_fun(x)
   } else {